eth-options-bot/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
websockets>=12.0
requests>=2.31.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
pyarrow>=14.0.0
typer>=0.9.0
pyyaml>=6.0.0
pytest>=7.4.0
matplotlib>=3.7.0
